pyproject.toml
[tool.poetry]
name = "cyclingsimilarity-com"
version = "0.1.0"
description = "Backbone repository for the cyclingsimilarity.com project"
authors = ["Samuel Borms"]
license = "MIT"
readme = "README.md"
packages = [{include = "cyclingsimilarity"}]

[tool.poetry.dependencies]
python = ">3.9.7,<4.0"
black = "^23.9.1"
isort = "^5.12.0"
ruff = "^0.0.290"
pre-commit = "^3.4.0"
numpy = "^1.25.2"
pandas = "^2.1.0"
torch = "^2.0.1"
fastai = "^2.7.12"
scikit-learn = "^1.3.0"
procyclingstats = "0.1.7"
jupyter = "^1.0.0"
pytest = "^7.4.2"
ipykernel = "^6.25.2"
openpyxl = "^3.1.2"
pytest-cov = "^4.1.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
streamlit = "^1.26.0"
fire = "^0.5.0"
faiss-cpu = "^1.7.4"
boto3 = "^1.28.50"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
